some bugs fixed

+++ SymbolTable/2105081_hash.hpp
@@ -5,9 +5,9 @@ using namespace std;
 class Hash
 {
     public:
-        static uint64_t sdbm_hash(string key_string)
+        static unsigned int sdbm_hash(string key_string)
 	    {
-	    	uint64_t hash_code=0;
+	    	unsigned int hash_code=0;
 	    	int len=key_string.length();
 	    	for(int i=0;i<len;i++)
 	    	{
